+++ Project.Core/Interfaces/IMapper/IBaseMapper.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Project.Core.Interfaces.IMapper
+{
+    public interface IBaseMapper<TSource, TDestination>
+    {
+        TDestination MapModel(TSource source);
+        IEnumerable<TDestination> MapList(IEnumerable<TSource> source);
+    }
+}
