+++ Project.Core/Interfaces/IRepositories/IProductRepository.cs
@@ -1,0 +1,13 @@
+﻿using Project.Core.Entities.General;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Project.Core.Interfaces.IRepositories
+{
+    public interface IProductRepository : IBaseRepository<Product>
+    {
+    }
+}
