+++ Project.Core/Project.Core.csproj
