+++ Project.Infrastructure/Repositories/ProductRepository.cs
@@ -1,0 +1,19 @@
+﻿using Project.Core.Entities.General;
+using Project.Core.Interfaces.IRepositories;
+using Project.Infrastructure.Data;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Project.Infrastructure.Repositories
+{
+    public class ProductRepository : BaseRepository<Product>, IProductRepository
+    {
+        public ProductRepository(ApplicationDbContext dbContext) : base(dbContext)
+        {
+        }
+
+    }
+}
