+++ Project.UI/Controllers/ProductController.cs
@@ -1,0 +1,103 @@
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Project.Core.Interfaces.IServices;
+
+namespace Project.UI.Controllers
+{
+    public class ProductController : Controller
+    {
+        private readonly ILogger<ProductController> _logger;
+        private readonly IProductService _productService;
+
+        public ProductController(ILogger<ProductController> logger, IProductService productService)
+        {
+            _logger = logger;
+            _productService = productService;
+        }
+
+        // GET: ProductController
+        public async Task<IActionResult> Index()
+        {
+            try
+            {
+                var products = await _productService.GetProducts();
+                return View(products);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "An error occurred while retrieving products");
+                return StatusCode(500, ex.Message);
+            }
+
+        }
+
+        // GET: ProductController/Details/5
+        public ActionResult Details(int id)
+        {
+            return View();
+        }
+
+        // GET: ProductController/Create
+        public ActionResult Create()
+        {
+            return View();
+        }
+
+        // POST: ProductController/Create
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Create(IFormCollection collection)
+        {
+            try
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            catch
+            {
+                return View();
+            }
+        }
+
+        // GET: ProductController/Edit/5
+        public ActionResult Edit(int id)
+        {
+            return View();
+        }
+
+        // POST: ProductController/Edit/5
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Edit(int id, IFormCollection collection)
+        {
+            try
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            catch
+            {
+                return View();
+            }
+        }
+
+        // GET: ProductController/Delete/5
+        public ActionResult Delete(int id)
+        {
+            return View();
+        }
+
+        // POST: ProductController/Delete/5
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Delete(int id, IFormCollection collection)
+        {
+            try
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            catch
+            {
+                return View();
+            }
+        }
+    }
+}
