+++ Project.UI/Extensions/ServiceExtension.cs
@@ -1,0 +1,45 @@
+﻿using AutoMapper;
+using Microsoft.Extensions.DependencyInjection;
+using Project.Core.Entities.General;
+using Project.Core.Interfaces.IMapper;
+using Project.Core.Interfaces.IRepositories;
+using Project.Core.Interfaces.IServices;
+using Project.Core.Mapper;
+using Project.Core.Services;
+using Project.Infrastructure.Repositories;
+
+namespace Project.UI.Extensions
+{
+    public static class ServiceExtension
+    {
+        public static IServiceCollection RegisterService(this IServiceCollection services)
+        {
+            #region Services
+            services.AddScoped<IProductService, ProductService>();
+
+            #endregion
+
+            #region Repositories
+            services.AddTransient<IProductRepository, ProductRepository>();
+
+            #endregion
+
+            #region Mapper
+            var configuration = new MapperConfiguration(cfg =>
+            {
+                cfg.CreateMap<Product, ProductViewModel>();
+                cfg.CreateMap<ProductViewModel, Product>();
+            });
+
+            IMapper mapper = configuration.CreateMapper();
+
+            // Register the IMapperService implementation with your dependency injection container
+            services.AddSingleton<IBaseMapper<Product, ProductViewModel>>(new BaseMapper<Product, ProductViewModel>(mapper));
+            services.AddSingleton<IBaseMapper<ProductViewModel, Product>>(new BaseMapper<ProductViewModel, Product>(mapper));
+
+            #endregion
+
+            return services;
+        }
+    }
+}
