+++ Project.UI/Views/Home/Index.cshtml
@@ -1,0 +1,7 @@
+﻿@{
+    ViewData["Title"] = "Index";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+<div class="container">
+    <h3>Home Page</h3>
+</div>
