+++ Project.UI/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using Project.Core.Entities.General;
+@using Project.UI
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
