--- ProjectMaster.Core/Entities/General/Base.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProjectMaster.Core.Entities.General
-{
-    public class Base<T>
-    {
-        [Key]
-        public T Id { get; set; }
-        public DateTime? EntryDate { get; set; }
-        public DateTime? UpdateDate { get; set; }
-    }
-}
