--- ProjectMaster.Core/Interfaces/IRepositories/IProductRepository.cs
@@ -1,13 +1,0 @@
-﻿using ProjectMaster.Core.Entities.General;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProjectMaster.Core.Interfaces.IRepositories
-{
-    public interface IProductRepository : IBaseRepository<Product>
-    {
-    }
-}
