--- ProjectMaster.Core/Mapper/BaseMapper.cs
@@ -1,30 +1,0 @@
-﻿using AutoMapper;
-using ProjectMaster.Core.Interfaces.IMapper;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProjectMaster.Core.Mapper
-{
-    public class BaseMapper<TSource, TDestination> : IBaseMapper<TSource, TDestination>
-    {
-        private readonly IMapper _mapper;
-
-        public BaseMapper(IMapper mapper)
-        {
-            _mapper = mapper;
-        }
-
-        public TDestination MapModel(TSource source)
-        {
-            return _mapper.Map<TDestination>(source);
-        }
-
-        public IEnumerable<TDestination> MapList(IEnumerable<TSource> source)
-        {
-            return _mapper.Map<IEnumerable<TDestination>>(source);
-        }
-    }
-}
