--- ProjectMaster.Infrastructure/Repositories/BaseRepository.cs
@@ -1,59 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using ProjectMaster.Core.Exceptions;
-using ProjectMaster.Core.Interfaces.IRepositories;
-using ProjectMaster.Infrastructure.Data;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProjectMaster.Infrastructure.Repositories
-{
-    public class BaseRepository<T> : IBaseRepository<T> where T : class
-    {
-        private readonly ApplicationDbContext _dbContext;
-
-        public BaseRepository(ApplicationDbContext dbContext)
-        {
-            _dbContext = dbContext;
-        }
-
-        public async Task<IEnumerable<T>> GetAll()
-        {
-            return await _dbContext.Set<T>().ToListAsync();
-        }
-
-        public async Task<T> GetById<Tid>(Tid id)
-        {
-            var data = await _dbContext.Set<T>().FindAsync(id);
-            if (data == null)
-                throw new NotFoundException("No data found");
-            return data;
-        }
-
-        public async Task<T> Create(T model)
-        {
-            await _dbContext.Set<T>().AddAsync(model);
-            await _dbContext.SaveChangesAsync();
-            return model;
-        }
-
-        public async Task Update(T model)
-        {
-            _dbContext.Set<T>().Update(model);
-            await _dbContext.SaveChangesAsync();
-        }
-
-        public async Task Delete(T model)
-        {
-            _dbContext.Set<T>().Remove(model);
-            await _dbContext.SaveChangesAsync();
-        }
-
-        public async Task SaveChangeAsync()
-        {
-            await _dbContext.SaveChangesAsync();
-        }
-    }
-}
