--- ProjectMaster.Infrastructure/Repositories/ProductRepository.cs
@@ -1,19 +1,0 @@
-﻿using ProjectMaster.Core.Entities.General;
-using ProjectMaster.Core.Interfaces.IRepositories;
-using ProjectMaster.Infrastructure.Data;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProjectMaster.Infrastructure.Repositories
-{
-    public class ProductRepository : BaseRepository<Product>, IProductRepository
-    {
-        public ProductRepository(ApplicationDbContext dbContext) : base(dbContext)
-        {
-        }
-
-    }
-}
