--- ProjectMaster.UI/Controllers/HomeController.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace ProjectMaster.UI.Controllers
-{
-    public class HomeController : Controller
-    {
-        public IActionResult Index()
-        {
-            return View();
-        }
-    }
-}
