--- ProjectMaster.UI/Program.cs
@@ -1,30 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using ProjectMaster.Infrastructure.Data;
-using ProjectMaster.UI.Extensions;
-
-var builder = WebApplication.CreateBuilder(args);
-
-builder.Services.AddDbContext<ApplicationDbContext>(options =>
-{
-    options.UseSqlServer(builder.Configuration.GetConnectionString("PrimaryDbConnection"));
-});
-
-builder.Services.AddControllersWithViews();
-builder.Services.RegisterService();
-
-
-var app = builder.Build();
-
-if (app.Environment.IsDevelopment())
-{
-    app.UseDeveloperExceptionPage();
-}
-
-app.UseRouting();
-
-app.UseEndpoints(endpoints =>
-{
-    endpoints.MapDefaultControllerRoute();
-});
-
-app.Run();
