--- ProjectMaster.UI/Views/Home/Index.cshtml
@@ -1,7 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Index";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-<div class="container">
-    <h3>Index</h3>
-</div>
