--- ProjectMaster.UI/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using ProjectMaster.Core.Entities.General;
-@using ProjectMaster
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
