--- ProjectMaster.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.6.33712.159
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProjectMaster.Core", "ProjectMaster.Core\ProjectMaster.Core.csproj", "{540CC202-D687-4B54-8746-4E943F0D16C6}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProjectMaster.Infrastructure", "ProjectMaster.Infrastructure\ProjectMaster.Infrastructure.csproj", "{A9767277-DD3A-47BA-9704-E525D32F1D9D}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProjectMaster.UI", "ProjectMaster.UI\ProjectMaster.UI.csproj", "{CAA1C682-F635-4957-B4F2-01680FFD1A14}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{540CC202-D687-4B54-8746-4E943F0D16C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{540CC202-D687-4B54-8746-4E943F0D16C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{540CC202-D687-4B54-8746-4E943F0D16C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{540CC202-D687-4B54-8746-4E943F0D16C6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A9767277-DD3A-47BA-9704-E525D32F1D9D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A9767277-DD3A-47BA-9704-E525D32F1D9D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A9767277-DD3A-47BA-9704-E525D32F1D9D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A9767277-DD3A-47BA-9704-E525D32F1D9D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CAA1C682-F635-4957-B4F2-01680FFD1A14}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CAA1C682-F635-4957-B4F2-01680FFD1A14}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CAA1C682-F635-4957-B4F2-01680FFD1A14}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CAA1C682-F635-4957-B4F2-01680FFD1A14}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {554B3771-ADE1-4A3F-8808-562B64E046E0}
-	EndGlobalSection
-EndGlobal
