--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "bootstrap-icons": "^1.10.5"
-  }
-}
